// src/app/core/enums/validation-constraint.enum.ts
export enum ValidationConstraint {
  IS_NOT_EMPTY = 'isNotEmpty',
  IS_STRING = 'isString',
  IS_INT = 'isInt',
  IS_EMAIL = 'isEmail',
  IS_NUMBER = 'isNumber',
  IS_DATE_STRING = 'isDateString',
  IS_UUID = 'isUuid',
  IS_POSITIVE = 'isPositive',
  LENGTH = 'length',
  MAX_LENGTH = 'maxLength',
  MIN_LENGTH = 'minLength',
  DATE = 'date',
  MIN = 'min',
  MAX = 'max',
  ARRAY_MIN_SIZE = 'arrayMinSize',
  ARRAY_MAX_SIZE = 'arrayMaxSize',
  IS_DEFINED = 'isDefined',
}